backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.20.0
pydantic==2.7.0
sqlalchemy[asyncio]==2.0.44
asyncpg==0.29.0
alembic==1.13.2
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0
spacy==3.6.0
aiofiles==23.2.1
